Workaround for texels shifting about between alternate frames (on walls and floors). Hack/temp fix for precision issues in the rasterizer when alternating between the 2 framebuffers used by DOOM - switching to a different framebuffer seemed to produce different rasterization results.
@JaCzekanski just to let you know I'm seeing some strange flickering/shifting on some textures in DOOM with the latest code. This hack (using double precision) seems to alleviate the issue for me in the meantime.
I also can't boot the original ROM in the latest versions of Avocado to tell which commit the issues started appearing in. The game appears to freeze before reaching the main menu - perhaps something CD ROM related?
PsyDoom is still fine however, since it has removed or simplified a lot of the CD ROM emulation.